pyproject.toml
[project]
name = "cj12"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["starlette>=0.47.2", "uvicorn>=0.35.0"]

[dependency-groups]
dev = [
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "pre-commit>=4.2.0",
    "pyodide-py>=0.28.1",
    "pytest>=8.4.1",
]

[tool.ruff]
lint.select = ["ALL"]
lint.ignore = ["D", "ANN401", "TD002", "TD003", "FIX002"]
lint.per-file-ignores = { "tests/**/*.py" = ["S101", "PLR2004"] }

[tool.pyright]
typeCheckingMode = "off"
